Fix issue 87
Let user to set its own configuration when configuring
TranslationOptions for the Page model.
User has to unregister Page and register it again with the new
configuration.
For example the user can now set title and slug fields required for all
languages for any model that subclasses the Page model.